+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/abstracts/AbstractMutationOperator.java
@@ -9,7 +9,7 @@ import hr.dlatecki.algorithms.gen_alg.population.interfaces.IChromosome;
  * class instead of directly implementing <code>IMutationOperator</code> interface.
  * 
  * @author Domagoj Latečki
- * @version 1.1
+ * @version 1.0
  * @since 1.8
  * @param <C> the type of chromosome which will be used in the mutation operator.
  * @see IChromosome