+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/interfaces/ISelectionOperator.java
@@ -9,7 +9,7 @@ import hr.dlatecki.algorithms.gen_alg.population.interfaces.IChromosome;
  * operator extend <code>AbstractSelectionOperator</code> instead of directly implementing this interface.
  * 
  * @author Domagoj Latečki
- * @version 1.1
+ * @version 1.0
  * @since 1.8
  * @param <C> Type of chromosome which will be used in the selection operator.
  * @see IChromosome