+++ src/main/java/hr/dlatecki/algorithms/gen_alg/population/chromosomes/ByteArrayWrapperChromosome.java
@@ -1,32 +1,32 @@
 package hr.dlatecki.algorithms.gen_alg.population.chromosomes;
 
 import java.util.Arrays;
-import hr.dlatecki.algorithms.gen_alg.codecs.interfaces.IDoubleArrayCodec;
+import hr.dlatecki.algorithms.gen_alg.codecs.interfaces.IByteArrayCodec;
 import hr.dlatecki.algorithms.gen_alg.population.abstracts.AbstractChromosome;
 
 /**
- * An implementation of the chromosome which contains an array of <code>double</code>s and a reference to the wrapped
+ * An implementation of the chromosome which contains an array of <code>byte</code>s and a reference to the wrapped
  * object. The wrapped object is represented by the array. This class also provides methods for fetching the wrapped
  * object and an array which represents it. When a change occurs the array, a new object is wrapped which corresponds to
  * the array, and vice-versa. Essentially, the wrapped object is always represented by the array which was encoded from
  * it, and the array can always be decoded into an object with same attributes. Encoding and decoding process is
  * performed by the coder which is provided upon creation of the instance of this class.<br>
  * <br>
- * To construct a chromosome using an item, use {@link #fromMutable(Object, IDoubleArrayCodec)} or
- * {@link #fromImmutable(Object, IDoubleArrayCodec)} static methods of this class.
+ * To construct a chromosome using an item, use {@link #fromMutable(Object, IByteArrayCodec)} or
+ * {@link #fromImmutable(Object, IByteArrayCodec)} static methods of this class.
  * 
  * @author Domagoj Latečki
  * @version 1.0
  * @param <I> type of the object to wrap.
  * @since 1.8
- * @see DoubleArrayChromosome
+ * @see ByteArrayChromosome
  */
-public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
+public class ByteArrayWrapperChromosome<I> extends ByteArrayChromosome {
     
     /**
      * Serial version UID.
      */
-    private static final long serialVersionUID = 6889342943515258765L;
+    private static final long serialVersionUID = 1351582272510994412L;
     /**
      * Indicates if the currently stored item is immutable.
      */
@@ -42,12 +42,12 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
     /**
      * Codec which is used in item encoding/decoding.
      */
-    private IDoubleArrayCodec<I> codec;
+    private IByteArrayCodec<I> codec;
     
     /**
-     * Constructs a <code>DoubleArrayWrapperChromosome</code> object using the provided mutable item and codec. If a
-     * chromosome for immutable item is needed, use {@link #fromImmutable(Object, IDoubleArrayCodec)} instead. The item
-     * will be encoded into the array of <code>double</code>s which will represent the encoded item. The encoding action
+     * Constructs a <code>ByteArrayWrapperChromosome</code> object using the provided mutable item and codec. If a
+     * chromosome for immutable item is needed, use {@link #fromImmutable(Object, IByteArrayCodec)} instead. The item
+     * will be encoded into the array of <code>byte</code>s which will represent the encoded item. The encoding action
      * is performed by the provided codec. New item will be created and stored by decoding an array which was just
      * encoded from the provided item. This will ensure that the item wrapped in this object cannot be changed from
      * outside of this object.
@@ -56,28 +56,28 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
      * @param codec codec to use in item encoding/decoding process.
      * @return Constructed chromosome.
      */
-    public static <I> DoubleArrayWrapperChromosome<I> fromMutable(I item, IDoubleArrayCodec<I> codec) {
+    public static <I> ByteArrayWrapperChromosome<I> fromMutable(I item, IByteArrayCodec<I> codec) {
         
-        return new DoubleArrayWrapperChromosome<>(item, false, codec);
+        return new ByteArrayWrapperChromosome<>(item, false, codec);
     }
     
     /**
-     * Constructs a <code>DoubleArrayWrapperChromosome</code> object using the provided immutable item and codec. If a
-     * chromosome for mutable item is needed, use {@link #fromMutable(Object, IDoubleArrayCodec)} instead. The item will
-     * be encoded into the array of <code>double</code>s which will represent the encoded item. The encoding action is
+     * Constructs a <code>ByteArrayWrapperChromosome</code> object using the provided immutable item and codec. If a
+     * chromosome for mutable item is needed, use {@link #fromMutable(Object, IByteArrayCodec)} instead. The item will
+     * be encoded into the array of <code>byte</code>s which will represent the encoded item. The encoding action is
      * performed by the provided codec. The item will be wrapped in this object.
      * 
      * @param item item to wrap in the chromosome object.
      * @param codec codec to use in item encoding/decoding process.
      * @return Constructed chromosome.
      */
-    public static <I> DoubleArrayWrapperChromosome<I> fromImmutable(I item, IDoubleArrayCodec<I> codec) {
+    public static <I> ByteArrayWrapperChromosome<I> fromImmutable(I item, IByteArrayCodec<I> codec) {
         
-        return new DoubleArrayWrapperChromosome<I>(item, true, codec);
+        return new ByteArrayWrapperChromosome<I>(item, true, codec);
     }
     
     /**
-     * Constructs a <code>DoubleArrayWrapperChromosome</code> object with provided codec. If the item which will be
+     * Constructs a <code>ByteArrayWrapperChromosome</code> object with provided codec. If the item which will be
      * wrapped in this object is meant to be immutable, then set the <code>immutable</code> flag to <code>true</code>.
      * When flag is set to <code>true</code> the provided item will be set as the wrapped item. If the flag is set to
      * <code>false</code>, then the new item will be created by decoding an array which was encoded from the provided
@@ -86,14 +86,14 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
      * @param immutable indicates if item which will be stored in this object is immutable.
      * @param codec codec to use in item encoding/decoding process.
      */
-    public DoubleArrayWrapperChromosome(boolean immutable, IDoubleArrayCodec<I> codec) {
+    public ByteArrayWrapperChromosome(boolean immutable, IByteArrayCodec<I> codec) {
         this.immutable = initialImmutable = immutable;
         this.codec = codec;
     }
     
     /**
-     * Constructs a <code>DoubleArrayWrapperChromosome</code> object using the provided item and codec. The item will be
-     * encoded into the array of <code>double</code>s which will represent the encoded item. The encoding action is
+     * Constructs a <code>ByteArrayWrapperChromosome</code> object using the provided item and codec. The item will be
+     * encoded into the array of <code>byte</code>s which will represent the encoded item. The encoding action is
      * performed by the provided codec. If item is immutable, then set the <code>immutable</code> flag to
      * <code>true</code>. When flag is set to <code>true</code> the provided item will be set as the wrapped item. If
      * the flag is set to <code>false</code>, then the new item will be created by decoding an array which was encoded
@@ -104,31 +104,31 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
      * @param immutable indicates if <code>item</code> is immutable.
      * @param codec codec to use in item encoding/decoding process.
      */
-    private DoubleArrayWrapperChromosome(I item, boolean immutable, IDoubleArrayCodec<I> codec) {
+    private ByteArrayWrapperChromosome(I item, boolean immutable, IByteArrayCodec<I> codec) {
         this(immutable, codec);
         
-        values = codec.encode(item);
+        bytes = codec.encode(item);
         
-        this.item = immutable ? item : codec.decode(values);
+        this.item = immutable ? item : codec.decode(bytes);
     }
     
     /**
-     * Constructs a <code>DoubleArrayWrapperChromosome</code> object using the provided array of <code>double</code>s
-     * and codec. The array will be copied into this object, so outside changes to the array will have no effect on the
+     * Constructs a <code>ByteArrayWrapperChromosome</code> object using the provided array of <code>byte</code>s and
+     * codec. The array will be copied into this object, so outside changes to the array will have no effect on the
      * contents of this object. It will also be decoded into an item by the codec. The item will be wrapped in this
      * object.<br>
      * <br>
-     * To construct a chromosome using an item, use {@link #fromMutable(Object, IDoubleArrayCodec)} or
-     * {@link #fromImmutable(Object, IDoubleArrayCodec)} static method of this class.
+     * To construct a chromosome using an item, use {@link #fromMutable(Object, IByteArrayCodec)} or
+     * {@link #fromImmutable(Object, IByteArrayCodec)} static method of this class.
      * 
-     * @param values values of the chromosome.
+     * @param bytes bytes of the chromosome.
      * @param codec codec to use in item encoding/decoding process.
      */
-    public DoubleArrayWrapperChromosome(double[] values, IDoubleArrayCodec<I> codec) {
+    public ByteArrayWrapperChromosome(byte[] bytes, IByteArrayCodec<I> codec) {
         this(true, codec);
-        this.values = Arrays.copyOf(values, values.length);
+        this.bytes = Arrays.copyOf(bytes, bytes.length);
         
-        item = codec.decode(this.values);
+        item = codec.decode(this.bytes);
     }
     
     @Override
@@ -136,15 +136,15 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
     protected void deepCopyTo(AbstractChromosome target) {
         
         // Type is guaranteed to be DoubleArrayWrapperChromosome<I> because the target is cloned from this object.
-        ((DoubleArrayWrapperChromosome<I>) target).item = immutable ? item : codec.decode(values);
+        ((ByteArrayWrapperChromosome<I>) target).item = immutable ? item : codec.decode(bytes);
     }
     
     @Override
-    public void setValues(double[] values) {
+    public void setBytes(byte[] bytes) {
         
-        super.setValues(values);
+        super.setBytes(bytes);
         
-        item = codec.decode(this.values);
+        item = codec.decode(this.bytes);
         immutable = true;
     }
     
@@ -156,7 +156,7 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
      */
     public I getItem() {
         
-        return immutable ? item : codec.decode(values);
+        return immutable ? item : codec.decode(bytes);
     }
     
     /**
@@ -168,8 +168,8 @@ public class DoubleArrayWrapperChromosome<I> extends DoubleArrayChromosome {
     public void setItem(I item) {
         
         immutable = initialImmutable;
-        values = codec.encode(item);
+        bytes = codec.encode(item);
         
-        this.item = immutable ? item : codec.decode(values);
+        this.item = immutable ? item : codec.decode(bytes);
     }
 }